Refactor hab prompt into separate module
