+++ firmware/NOKHAB/hab-server/hab_server.py
@@ -3,67 +3,15 @@ import socket
 import sys
 import threading
 
-from cmdr import Cmdr
+from hab_prompt import HabPrompt
 from sensor import Sensor, SensorCmd
+from threading import Thread
 
 HOST = '0.0.0.0'
 PORT = 1234
 
 sensor: Sensor = Sensor()
 sensor_cmd: SensorCmd = SensorCmd(sensor)
-
-class HabPrompt(Cmdr):
-    prompt = 'hab> '
-
-    def __init__(self, sock:socket.socket):
-        self._sock = sock
-        sock_file = sock.makefile('rw')
-        Cmdr.__init__(self, stdin=sock_file, stdout=sock_file)
-        sys.stdout = sock_file
-        self.use_rawinput = False
-
-    def do_exit(self, args):
-        """ exit HAB shell """
-        print("Thank you for flying HAB airlines", flush=True)
-        self._sock.shutdown(socket.SHUT_RDWR)
-        self._sock.close()
-
-    def preloop(self):
-        print('''
-                    ,~-.
-                   (  ' )-.          ,~'`-.
-                ,~' `  ' ) )       _(   _) )
-               ( ( .--.===.--.    (  `    ' )
-                `.%%.;::|888.#`.   `-'`~~=~'
-                /%%/::::|8888\##\\
-               |%%/:::::|88888\##|
-               |%%|:::::|88888|##|.,-.
-               \%%|:::::|88888|##/    )_
-                \%\:::::|88888/#/ ( `'  )
-                 \%\::::|8888/#/(  ,  -'`-.
-             ,~-. `%\:::|888/#'(  (     ') )
-             (  ) )_ `\__|__/'   `~-~=--~~='
-            ( ` ')  ) [VVVVV]
-           (_(_.~~~'   \|_|/
-                       [XXX]
-                       `"""' 
-        ''')
-        print("Welcome to the HAB shell")
-
-    def postloop(self):
-        print("Thank you for flying HAB airlines")
-
-def service_connection(sock:socket.socket, address):
-    try:
-        print(address)
-        p = HabPrompt(sock)
-        p.register(sensor_cmd)
-        p.cmdloop()
-    except:
-        pass
-    finally:
-        sock.close()
-        return False
 
 def accept_connection():
     s = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
@@ -72,13 +20,13 @@ def accept_connection():
     s.listen(5)
     while True:
         client, address = s.accept()
-        _, port = address
         client.settimeout(None)
-        threading.Thread(target=service_connection, args=(client, address), name='service_connection_'+str(port), daemon=True).start()
+        p = HabPrompt(client)
+        p.register(sensor_cmd)
+        p.start()
 
 if __name__ == "__main__":
-    thread = threading.Thread(target=accept_connection, name='accept_connection')
-    thread.daemon = True
+    thread = threading.Thread(target=accept_connection, name='accept_connection', daemon=False)
     thread.start()
     sensor.start()
     try:
